Add --idle_timeout flag to exit bazel-remote when idling for certain time
This is for #61.
The max_idle_time flag is useful when using bazel-remote as a local proxy
to PUT cache objects to remote cache server and exit when idle for a time
specified by max_idle_time flag. This allows for local build finish early
and leave bazel-remote to finish the PUT request queue, instead of
waiting for the PUT requests to finish.