disk cache tests: return errors from helpers, fail at the call site (#108)
If we call t.Fatal inside test helpers, then the log message says that
the test failed inside that function. But what we really want to know
is where the helper was called from. So return errors and call t.Fatal
from the call site.

This increases the amount of boilerplate around these helpers, but
gives better error messages.

The tempDir function has no test-specific failures, and can be excluded.

Cleanup before implementing #106.