validate ActionCache items before returning them for http also (#111)
A common bug is the scenario where an ActionResult exists in the
ActionCache and is returned to the client, but not all the blobs
that it depends on are available in the CAS when the client tries
to get them. The client typically errors out at this point, and
might also continue failing when re-run, until the ActionResult
is removed from the cache.

To avoid this problem, we should verify that all the required
blobs are available before returning a given ActionCache. Doing
so should move these items to the "safe" end of the LRU queue,
and increase the odds that the CAS blobs will still exist in the
cache when the client requests them.

We previously had this check in the gRPC interface, but not for
the REST interface. This change refactors the check so it's
available to the REST interface. Since some clients might be
relying on bazel-remote to store arbitrary non-ActionResult blobs
in the ActionCache via the REST interface, we provide a command
line flag to disable this validation:
--disable_http_ac_validation

When --disable_http_ac_validation is specified, ActionCache blobs
uploaded and downloaded via REST are stored in a new "raw"
keyspace to avoid clashes with validated gRPC ActionResult blobs
(which are always validated and always stored in the "ac" keyspace).

If you want to use --disable_http_ac_validation and not lose access
to existing cache items, stop bazel-remote, rename the "ac" dir to
"raw", then restart bazel-remote with --disable_http_ac_validation.

Fixes #106.