ensure ExecutionMetadata.Worker is set on all uploaded ActionResults
proto3 does not encode default values, so if all fields in an ActionResult
are default then the serialized value is zero-length. We would like to
avoid using zero-length files, because this is the most common symptom of
filesystem corruption. Unfortunately, a default-valued ActionResult is
actually valid, so we can't reject it automatically.

We can modify ActionResult messages without changing their meaning, by
ensuring that one of the ExecutionMetadata's fields to a non-default
value. ExecutionMetadata.Worker seems like a good choice- the server can
set it to a meaningful value even if the client (or worker) left it empty.

This allows us to keep the zero-sized file corruption check when serving
ActionCache results.

Fixes #121.