add an optional http profiling endpoint
If --profile_port is specified (as a positive integer), then start a
http server on that port for runtime profiling data.

The host part of the address defaults to 127.0.0.1, and can be set
with the --profile_host flag. This is needed if you want to access
the profiling URLs from bazel-remote running inside docker.

Implements #154.