refactor the cache.Cache interface
Since all the proxy backends use the same copy+pasted logic to use
the disk cache, refactor to put that functionality in the disk cache.
This simplifies the proxy backends.

Contains is now also proxied to the backend, instead of only
returning whether or not the local disk cache contains the item.

Fixes #161.