disk cache: keep better track of currentSize when using a proxy backend
When using a proxy backend, on disk cache misss we add a placeholder
LRU entry with size -1, which incorrectly decreases the currentSize
counter by 1 (mistake 1).

Then if we get a proxy backend cache hit, the placeholder entry's size
field is updated directly, without updating the currentSize counter
(mistake 2). This causes a significant undercounting.

Otherwise if we get a proxy backend cache miss, then the first mistake
is undone.

So:
* Don't modify LRU item sizes directly, use lruItem.Add to update
  values.
* Specify a preliminary size of 0 when reserving a place in the
  index before attempting a proxy backend download.

Fixes #181.