disk cache: avoid leaking cache files in onEvict function
If there's an ongoing blob upload when that blob is evicted, then the
item was deleted from the LRU index, but the files on disk were left
behind and no longer included in the cache size accounting.

When we evict a blob that's being uploaded, we might need to remove:
1) Both the temp file and the regular cache file.
2) Just the regular cache file.

Note that if you're encountering this situation, it might be a sign
that your cache size is too small- when blob uploads start the item
is moved to the most-recently used end of the index, and the item
has moved all the way to the least-recently used end of the index
before the upload has completed.

Fixes #176.