automatically raise the limit on the number of open files
bazel-remote opens a lot of files, and can easily hit the limit unless
it is raised. Let's try to do that automatically.

This requires a workaround on macos, where getrlimit does not return
the correct hard limit for RLIMIT_NOFILE. We need to use the minimum
of the value from getrlimit and from sysctl. Since cgo is difficult
with bazel's rules_go just exec the sysctl command.

Fixes #174.

You might need to take additional steps to raise the hard limit,
depending on your operating system's defaults.