make Contains also return the size (or -1 if unknown)
In order to implement the remote asset API, we will need to check for
blobs and return a Digest. At the moment, our Contains function only
returns a boolean, but not the size of the item which is also required.

In most cases, we have the size information when Contains returns true,
but the value wasn't being plumbed through.

In case a proxy backend does not provide size information, return -1.