don't require explicit empty blob uploads for HTTP either
The gRPC API implementation has a hack to handle clients (like bazel?)
which in some cases don't upload the empty blob, but do reference it.
It is kind of OK not to upload the empty blob, since it can be recognised
by hash and trivially regenerated.

Let's imlpement the hack for the HTTP API also.

Implements #233.