simplify v0 directory migration and ignore irregular files
The original ("v0") directory structure placed all blobs directly
inside either the ac/ or cas/ directories, and we used filepath.walk
with a base case to skip the top directory and do a non-recursive
directory listing and move these files to the new ("v1") structure
with subdirectories based on the first two hex characters in the hash.

This base case only checked if ac/ or cas/ was a directory and failed if
it was a symlink. Let's fix that. It was also overcomplicated- we
should perform a simple directory listing instead of using filepath.Walk.

Fixes #285.