add some customized tempfile code
We are discussing allowing concurrent uploads for the same blob in #267.

Part of the proposed new design involves using pseudo-randomly named
temp files. Unfortunately, ioutil.TempFile creates files with more
restrictive permissions than we want- 0600 before umask, whereas we
want 0666 before umask.

This code is an adapted version of the algorithm in ioutil.TempFile,
refactored for our more narrow use case- in particular we skip the
use of the pid in the seed, since our use case is single-process,
and we don't bother re-seeding if we encounter lots of collisions
(which is unlikely).

(We do not use this code, yet.)