docker: simplify non-root user setup
It turns out that our base container image already has a "nonroot"
user, with uid/gid 65532.

The previous setup worked more of less out of the box when running as
uid 1000 (the default uid for the initial user on most linux setups).
Since we changed this default we need to document how to specify the
required user.

Followup to #302.