asset: drop known file extension check
Not all files have extensions (eg unix executables). This was being
reported to the client as NotFound, and bazel seems to have a bug
where it assumes that the result was always successful and then errors
out with a hash verification error.

We should be able to download the file in this particular case
regardless of the bazel bug.

Background: #310.