refactor disk.Cache to allow concurrent inserts
Instead of adding placeholder "uncommitted" items to the LRU, reserve space,
then write the incoming blob to a randomly named tempfile (and verify it if
possible). If everything went well, unreserve the space, add it to the LRU
index and move the tempfile to the final location.

This allows concurrent uploads (or proxy downloads) of the same blob, which
might be slightly wasteful but meets all our consistency requirements. We
can try to optimise this later, if needed.

The "commit" now happens in the main part of the function and the deferred
cleanup function is much simpler.

We no longer have the "uncommitted item evicted" issue to consider, but we
might run into cases where there are inserts that fail because there is too
much space reserved.

Fixes #267, #318.