[disk] make compressed cas blob files valid .zst
By using a zstd skippable frame for our header, the file becomes
a valid .zst file on disk. This makes it possible to inspect cache
files with standard tools.

The cost is only 8 bytes per (compressed) file. Compressed CAS blobs
are now limited to ~511TB (with 1M chunks), which should not be an
issue in practical use.