Add new endpoint_metrics_duration_buckets config option (#390)
This allows admins to configure a more suitable histogram range for their
deployment. If the option is not specified, the default remains the same
as the current values.

I don't think this is something that is a good idea to set via command line
arguments, so it's a config file option only for now.

Fixes #389.