[proxy] switch between v1 and v2 CAS blobs depending on the storage mode
Since we have two different on-disk CAS blob formats, and we don't want
to perform transcoding in the proxy backends, we need to use a different
path for v2 CAS blobs.

This means that for now, it's not possible to use storage_mode: zstd and
use the httpproxy backend to proxy blobs to another bazel-remote instance.

Also, when switching storage modes, stored blobs for the other storage
mode on the proxy backends will not be used.