[disk] rework the disk storage to avoid fs operations while holding the LRU lock
By adding a psuedo-random string to the filename for blobs and
storing that string in the index, we no longer need to rename
files while holding the LRU index lock.

To avoid incomplete files from being reused after restarting
bazel-remote, we use a bit of a hack: create the blob file with
the setgid bit set, and un-set it once the blob is complete.
On startup, remove any files with the setgid bit set.