[disk] pass the actual logical blob size to FileLocationBase
availableOrTryProxy returns -1 for foundSize if the blob is not available
in the disk cache. If we then get a hit from the proxy backend, we should
have the actual foundSize. We should then use that (not -1) when calling
FileLocationBase to construct the filename to store the blob in locally.

This might fix #420.