[grpc] fix variable shadowing in FetchBlob
In case cache.Contains doesn't return the correct size, i.e. -1, the
correct size needs to be set by getting the value and using the size
from the Get call.

Before this commit the size variable was not correctly updated because
the inner variable shadowed the outer one, this is now fixed.

Fixes #420