Fix empty zstd blob bystestream read infinite loop
If a client tries to download the empty blob via bytestream compressed
with zstd, we were leaking a goroutine stuck in a hot infinite loop.
rc.Read was returning 0, nil forever, and the the client would timeout.

This change fixes this bug, and adds a test that would hang if the
bug wasn't fixed.