[grpc] optimise FindMissingBlobs
FindMissingBlobs calls can contain lots of blob hashes to lookup. Instead
of locking and unlocking the index once per hash, we can perform the lookups
in batches. And instead of querying proxy backends sequentially, we can use
a pool of goroutines to make lots of concurrent queries.

For the first part to work, we need to do this in the disk package instead
of the server package (where the rest of the gRPC code is).

Relates to #408.