[disk] limit the number of concurrent file removals
The disk cache's eviction function removes files in a separate goroutine
for each, to avoid holding the lock while waiting for the filesystem.

If the disk/filesystem is too slow to keep up with the rate of file removals,
we can hit Go's default 10,000 operating system thread limits, and crash.
To avoid this, let's try limiting the number of concurrent file removals to
half the default limit (5,000).

This might fix #479.