[metrics] add lru eviction idle time metric
This adds a "now - atime" metric for the least recently used item in the
cache, which can be useful to identify undersized caches. It is updated
once per minute rather than on each eviction, in order to avoid holding
the index mutex while performing the filesystem stat operation.

Since this depends on filesystem atime values, the relatime mount option
will limit the resolution of this metric, and the noatime mount option
makes it useless.