[disk] optimise startup index sorting
Previously, the cache dir scanning code would return a slice of
a struct that included lruItem fields plus two extra fields: the
lookup key that a corresponding lruItem will be stored at, and
an os.FileInfo to use for atime sorting.

With this patch, the cache dir scanning code returns a scanResult
struct with a slice of lruItems (so they don't need to be created
later) and a corresponding slice of the lookup key and atime
metadata. Then we can sort this scanResult, use the slice of
lruItems and throw the metadata away.