Make idle_timeout setting also apply when using https
The main use-case for this flag is for communicating with bazel-remote
running on localhost, uploading asynchronously to a proxy backend, and
https support isn't particularly useful for this. But there's no reason
it couldn't work too, and I think it's better not to have a special
case to document (or have unsupported and undocumented).