s3: update timestamps in Contains() also
When the s3.update_timestamps option is enabled, we update the timestamp
of any blob after it is successfully retried from the backend. We should
do the same when checking for blob existence in the Contains() function.

I don't think this is sufficient to make builds-without-the-bytes work
with the s3 proxy backend when eviction is enabled, but it might help.

Relates to #624.