Make bazel-remote packages consumable as a module
The new import path is:
github.com/buchgr/bazel-remote/v2

Warning: bazel-remote go packages do not follow semantic versioning rules,
since it is primarily intended to be consumed as a standalone executable.

I aim to maintain backwards compatibility for the standalone executable
(within each major release), but provide no guarantees on the stability of
the packages used internally.

Relates to #635.