Fix kubernetes example
The bazel-remote binary flags falls back to env vars for default values,
meaning that `$BAZEL_REMOTE_MAX_SIZE` will be ignored if the binary was
invoked with `--max_size=5`.

The relased containers contains a default `--max_size` flag which the
README addresses by suggesting to pass your own `--max_size` flag when
running bazel-remote containerized.

But when that container as a k8s deployment, it's very usual to simply
configure everything through env vars. This creates a very hard to spot
scenario, where one might think that the max size is correctly
configured with the env var, but in reality the flag will take
precedence and run bazel-remote with the default 5 GiB size, regardless
of what the user configured.

This commit fixes this by correctly providing an `args: [...]` to
behave exactly as the user would expect it.