Optimize file removals (part 1)
This commit:

 - Performs evictions from a single background goroutine that receives
   files to be removed via a channel.

 - Throttles number of concurrent Put requests with semaphore (but not
   rejecting them).

In order to:

 - Avoid crashing on high load.

 - Achieve up to 3 times faster cache eviction.

 - Achieve up to 70% higher write throughput in scenario with many
   cache evictions.

The cache can grow above max_size when asynchronous files removals do
not catch up with new file writes. This is addressed in the following
part 2 commit. This issue was masqueraded in previous bazel-remote
versions by instead running out of operating system threads and crash.