Introduce disk_size_limit (part 2)
Introduce a disk_size_limit for the total disk space of:

 - Files currently in the cache.
 - Reserved space for files currently being uploaded.
 - Evicted files not yet removed.

Setting this limit is optional (at least for now).

Reservations for Put requests are rejected when
disk_size_limit is exceeded.

The prometheus gauge bazel_remote_disk_cache_size_bytes is
updated to be a max value for the previous 30 seconds,
in order to be aware of short spikes when tuning the
disk_size_limit configuration.

There is also a new prometheus gauge
bazel_remote_disk_cache_size_bytes_limit showing current
configured limits in order to help visualize if current size
is getting close to the limit and help tuning the
disk_size_limit.