Drop server support for TLS 1.0 and 1.1
At the time of writing, current go versions default to TLS 1.0 as the minimum
version when acting as a server (and 1.2 when acting as a client). This change
raises the minium version to 1.2 when acting as a server too.

If this causes trouble, we can consider adding a config flag to specify the
minimum TLS version.

Fixes #708.