prom: Add option to specify metrics prefix
It's useful to be able to prefix the metrics that are being reported
so as to more easily differentiate them from other applications that
are using the same metrics.

This change adds the necessary bits (environment var, config file,
command line) to support capturing the prefix and passing it through
to the httpmetrics library.