Add more human-readable GitDescribe item to the status page
bazel-remote has included a GitCommit field in the /status page for
some time, the git ref SHA1 value is not the most readable. By adding
a new GitDescribe field with the output from `git describe --tags`
it is easier to see which version of bazel-remote is running.

For details of the values this field can hold, see:
https://git-scm.com/docs/git-describe

Relates to #436.