Add more human-readable GitTags item to the status page
bazel-remote has included a GitCommit field in the /status page for
some time, the git ref SHA1 value is not the most readable. By adding
a new GitTags field with the comma-separated list of tags (if any)
for the current commit, it is easier to see which version of
bazel-remote is running.

Relates to #436.